+++ Assets/Scripts/server/Effects/SpeedBoost.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class SpeedBoost : Effect
 {
-    public SpeedBoost(int _duration, float _dmovementSpeed, int _priority, int _key)
+    public SpeedBoost(float _duration, float _dmovementSpeed, int _priority, int _key)
     {
         duration = _duration;
         dmovementSpeed = _dmovementSpeed * Time.deltaTime * 700;
